refactor: use single monitor CLI entry point and update docs
Drop codecarbon-monitor in favor of codecarbon monitor, add --log-level
there, and document warm hardware reuse plus the correct log-level default.
