+++ pyproject.toml
@@ -124,7 +124,6 @@ viz-legacy = [
 [project.scripts]
 carbonboard = "codecarbon.viz.carbonboard:main"
 codecarbon = "codecarbon.cli.main:main"
-codecarbon-monitor = "codecarbon.cli.monitor_main:main"
 
 [tool.taskipy.tasks]
 pre-commit-install = { cmd = "pre-commit install", help = "Install pre-commit hook." }
